+++ internal/provider/data_source_certificate.go
@@ -108,15 +108,24 @@ func dataSourceCertificate() *schema.Resource {
 							Computed:    true,
 							Description: "The SHA1 fingerprint of the public key of the certificate.",
 						},
+						"cert_pem": {
+							Type:     schema.TypeString,
+							Computed: true,
+							Description: "Certificate data in [PEM (RFC 1421)](https://datatracker.ietf.org/doc/html/rfc1421) format. " +
+								"**NOTE**: the [underlying](https://pkg.go.dev/encoding/pem#Encode) " +
+								"[libraries](https://pkg.go.dev/golang.org/x/crypto/ssh#MarshalAuthorizedKey) that generate this " +
+								"value append a `\\n` at the end of the PEM. " +
+								"In case this disrupts your use case, we recommend using " +
+								"[`trimspace()`](https://www.terraform.io/language/functions/trimspace).",
+						},
 					},
 				},
 				Description: "The certificates protecting the site, with the root of the chain first.",
 			},
 			"id": {
-				Type:     schema.TypeString,
-				Computed: true,
-				Description: "Unique identifier of this data source: " +
-					"randomly generated string (UTC time when data source was read).",
+				Type:        schema.TypeString,
+				Computed:    true,
+				Description: "Unique identifier of this data source: hashing of the certificates in the chain.",
 			},
 		},
 	}