+++ internal/provider/data_source_certificate_test.go
@@ -3,6 +3,7 @@ package provider
 import (
 	"fmt"
 	"regexp"
+	"strings"
 	"testing"
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
@@ -32,6 +33,8 @@ func TestAccDataSourceCertificate_CertificateContent(t *testing.T) {
 					resource.TestCheckResourceAttr("data.tls_certificate.test", "certificates.0.not_before", "2019-11-08T09:01:36Z"),
 					resource.TestCheckResourceAttr("data.tls_certificate.test", "certificates.0.not_after", "2019-11-08T19:01:36Z"),
 					resource.TestCheckResourceAttr("data.tls_certificate.test", "certificates.0.sha1_fingerprint", "61b65624427d75b61169100836904e44364df817"),
+					testCheckPEMFormat("data.tls_certificate.test", "certificates.0.cert_pem", PreambleCertificate),
+					resource.TestCheckResourceAttr("data.tls_certificate.test", "certificates.0.cert_pem", strings.TrimSpace(testTlsDataSourceCertFromContent)+"\n"),
 				),
 			},
 		},
@@ -374,6 +377,8 @@ func localTestCertificateChainCheckFunc() resource.TestCheckFunc {
 		resource.TestCheckResourceAttr("data.tls_certificate.test", "certificates.0.not_before", "2019-11-07T15:47:48Z"),
 		resource.TestCheckResourceAttr("data.tls_certificate.test", "certificates.0.not_after", "2019-12-17T15:47:48Z"),
 		resource.TestCheckResourceAttr("data.tls_certificate.test", "certificates.0.sha1_fingerprint", "5829a9bcc57f317719c5c98d1f48d6c9957cb44e"),
+		testCheckPEMFormat("data.tls_certificate.test", "certificates.0.cert_pem", PreambleCertificate),
+		resource.TestCheckResourceAttr("data.tls_certificate.test", "certificates.0.cert_pem", strings.TrimSpace(testTlsDataSourceCertFromURL00)+"\n"),
 
 		resource.TestCheckResourceAttr("data.tls_certificate.test", "certificates.1.signature_algorithm", "SHA256-RSA"),
 		resource.TestCheckResourceAttr("data.tls_certificate.test", "certificates.1.public_key_algorithm", "RSA"),
@@ -385,6 +390,8 @@ func localTestCertificateChainCheckFunc() resource.TestCheckFunc {
 		resource.TestCheckResourceAttr("data.tls_certificate.test", "certificates.1.not_before", "2019-11-08T09:01:36Z"),
 		resource.TestCheckResourceAttr("data.tls_certificate.test", "certificates.1.not_after", "2019-11-08T19:01:36Z"),
 		resource.TestCheckResourceAttr("data.tls_certificate.test", "certificates.1.sha1_fingerprint", "61b65624427d75b61169100836904e44364df817"),
+		testCheckPEMFormat("data.tls_certificate.test", "certificates.1.cert_pem", PreambleCertificate),
+		resource.TestCheckResourceAttr("data.tls_certificate.test", "certificates.1.cert_pem", strings.TrimSpace(testTlsDataSourceCertFromURL01)+"\n"),
 	)
 }